Assn 2-1. Make counter app & Divide components by file
2nd, divide components by file. The divided components are named CountApp, Counter, Buttons, Button.
