+++ src/Button.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Button({ onClick, children }) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+    >
+      {children}
+    </button>
+  );
+}
+
+export default Button;
