+++ src/Counter.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+function Counter({ count, onClick }) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+    >
+      Click me!
+      (
+      { count }
+      )
+    </button>
+  );
+}
+
+export default Counter;
